requirements.txt
aioredis==1.3.1
asgiref==3.2.3
async-timeout==3.0.1
attrs==19.3.0
autobahn==19.11.1
Automat==0.8.0
beautifulsoup4==4.8.2
boto3==1.12.16
botocore==1.15.16
bs4==0.0.1
certifi==2019.11.28
cffi==1.13.2
channels==2.4.0
channels-redis==2.4.1
chardet==3.0.4
constantly==15.1.0
countries==0.2.0
cryptography==2.8
currencies==2014.7.13
daphne==2.4.1
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.7
django-allauth==0.41.0
django-countries==5.5
django-crispy-forms==1.8.1
django-debug-toolbar==2.1
django-filter==2.2.0
django-heroku==0.3.1
django-storages==1.9.1
django-suit==0.2.28
docutils==0.15.2
future==0.18.2
gunicorn==20.0.4
hiredis==1.0.1
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
iso8601==0.1.12
jmespath==0.9.5
lxml==4.4.2
msgpack==0.6.2
oauthlib==3.1.0
optional-django==0.3.0
Pillow==6.2.1
psycopg2==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
PyHamcrest==1.9.0
pyOpenSSL==19.1.0
python-dateutil==2.8.1
python3-openid==3.1.0
pytils==0.3
pytz==2019.3
PyYAML==5.1.2
react==4.3.0
requests==2.22.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
selenium==3.141.0
serial==0.0.97
service-identity==18.1.0
six==1.13.0
soupsieve==1.9.5
sqlparse==0.3.0
stripe==2.41.1
suit==2.0.2
tools==0.1.9
Twisted==19.10.0
txaio==18.8.1
urllib3==1.25.7
whitenoise==5.0.1
zope.interface==4.7.1
